Rename graphql api -> graphql client
