+++ apps/product/src/components/product/Product.tsx
@@ -5,7 +5,7 @@ import CheckIcon from "../icons/CheckIcon";
 import HeartIcon from "../icons/HeartIcon";
 import Price from "../price/Price";
 import StockLevelLabel from "../stock-level-label/StockLevelLabel";
-import { graphQlQuery } from "../../../../../libs/shared/graphql-api";
+import { graphQlQuery } from "../../../../../libs/shared/graphql-client";
 import { ADD_ITEM_TO_ORDER } from "~/components/product/Product.graphql";
 import { cartQuantitiesChangedEventId, orderChangeEventId } from "../../../../../libs/shared/custom-events";
